Create FakeCardBrandFilter test
